requirements.txt
# ============================================================================
# DEPENDENCIES FOR ETL
# MongoDB -> PostgreSQL + Cassandra
# ============================================================================

# Database Drivers
pymongo>=4.6.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
cassandra-driver>=3.29.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.2
pycryptodome>=3.20.0

# Development (optional)
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.1.0
flake8>=7.0.0

